refactor(Services): :recycle: Move sankey chart creation to the address service
