+++ src/pages/Addresses/components/RequestsByAddress.tsx
@@ -21,11 +21,11 @@ import FlowPairsTable from '@pages/shared/FlowPair/FlowPairsTable';
 import Metrics from '@pages/shared/Metrics';
 import { SelectedFilters } from '@pages/shared/Metrics/Metrics.interfaces';
 import { FlowPairsResponse, RequestOptions } from 'API/REST.interfaces';
-import { VarColors } from 'colors';
 
 import { TAB_0_KEY, TAB_1_KEY, TAB_2_KEY, httpColumns } from '../Addresses.constants';
 import { RequestLabels, AddressesLabels, FlowPairsLabels } from '../Addresses.enum';
 import { RequestsByAddressProps } from '../Addresses.interfaces';
+import { AddressesController } from '../services';
 import { QueriesServices } from '../services/services.enum';
 
 const PREFIX_DISPLAY_INTERVAL_CACHE_KEY = 'service-display-interval';
@@ -133,26 +133,8 @@ const RequestsByAddress: FC<RequestsByAddressProps> = function ({ addressId, add
   const serverNamesId = servers.map(({ name }) => name).join('|');
   const startTime = servers.reduce((acc, process) => Math.min(acc, process.startTime), 0);
 
-  const sourceNodes =
-    servicePairs?.map(({ metric }) => ({
-      id: `${metric.sourceProcess || metric.sourceSite.split('@_@')[0]} (client)`,
-      nodeColor: metric.sourceProcess ? VarColors.Blue400 : undefined
-    })) || [];
-  const destNodes =
-    servicePairs?.map(({ metric }) => ({
-      id: metric.destProcess || metric.destSite.split('@_@')[0],
-      nodeColor: metric.destProcess ? VarColors.Blue400 : undefined
-    })) || [];
-
-  const nodes = [...sourceNodes, ...destNodes].filter((v, i, a) => a.findIndex((v2) => v2.id === v.id) === i);
-
   const enableMetric = servicePairMetricSelected !== 'none';
-  const links =
-    servicePairs?.map(({ metric, value }) => ({
-      source: `${metric.sourceProcess || metric.sourceSite.split('@_@')[0]} (client)`,
-      target: metric.destProcess || metric.destSite.split('@_@')[0],
-      value: enableMetric ? Number(value[1]) : 0.01
-    })) || [];
+  const { nodes, links } = AddressesController.getNodesAndLinksResources(servicePairs || [], enableMetric);
 
   return (
     <>
@@ -185,7 +167,7 @@ const RequestsByAddress: FC<RequestsByAddressProps> = function ({ addressId, add
 
       {viewSelected === TAB_2_KEY && (
         <Stack hasGutter>
-          {nodes.length && (
+          {!!nodes.length && (
             <StackItem>
               <Card>
                 <CardHeader>
